12.09. 근태관리 세션 추가, pk 타입 수정, 위젯 세부 기능 구현 중

+++ EASYWAY/src/main/java/org/easyway/domain/attendance/AttendanceVO.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class AttendanceVO implements Serializable {
 
-	private int attendanceId;
-	private int employeeId;
+	private Long attendanceId;
+	private Long employeeId;
 	private String attendanceStart; // 출근
 	private String attendanceOut; // 외근
 	private String attendanceEnd; // 퇴근
